requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-extensions==3.1.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
